+++ sequence/entanglement_management/__init__.py
@@ -1,4 +1,4 @@
-__all__ = ['entanglement_protocol', 'generation', 'purification', 'swapping']
+__all__ = ['entanglement_protocol', 'generation', 'swapping', 'purification']
 
 def __dir__():
     return sorted(__all__)