+++ sequence/topology/router_net_topo.py
@@ -4,11 +4,9 @@ from networkx import Graph, dijkstra_path, exception
 
 from .topology import Topology as Topo
 from ..kernel.timeline import Timeline
-from ..kernel.quantum_manager import KET_STATE_FORMALISM
+from ..kernel.quantum_manager import KET_STATE_FORMALISM, QuantumManager
 from .node import BSMNode, QuantumRouter
 from ..constants import SPEED_OF_LIGHT
-from ..kernel.quantum_manager import BELL_DIAGONAL_STATE_FORMALISM
-from ..config import USER_CONFIG_PATH
 
 
 class RouterNetTopo(Topo):
@@ -44,11 +42,6 @@ class RouterNetTopo(Topo):
     def __init__(self, conf_file_name: str):
         self.bsm_to_router_map = {}
         self.encoding_type = None
-
-        # you can pass the config file as an environment variable or as an argument to RouterNetTopo. If not using 
-        # RouterNetTopo directly or, if you're using a custom backend, pass the config as environment variable.  
-        if USER_CONFIG_PATH is not None:
-            conf_file_name = USER_CONFIG_PATH
         super().__init__(conf_file_name)
 
     def _load(self, filename: str):
@@ -69,13 +62,14 @@ class RouterNetTopo(Topo):
         self._generate_forwarding_table(config)
 
     def _add_timeline(self, config: dict):
-        stop_time = config.get(Topo.STOP_TIME, float('inf'))
+        stop_time = config.get(Topo.STOP_TIME, 10 ** 23)
         formalism = config.get(Topo.FORMALISM, KET_STATE_FORMALISM)
         truncation = config.get(Topo.TRUNC, 1)
+        QuantumManager.set_global_manager_formalism(formalism)
         if config.get(self.IS_PARALLEL, False):
             raise Exception("Please install 'psequence' package for parallel simulations.")
         else:
-            self.tl = Timeline(stop_time, formalism, truncation)
+            self.tl = Timeline(stop_time=stop_time, truncation=truncation)
 
     def _map_bsm_routers(self, config):
         for qc in config[Topo.ALL_Q_CHANNEL]: